requirements.txt
kafka-python==2.0.2
websocket-client==1.7.0
pyspark==3.5.1
python-dotenv==1.0.1